Report_automation
This VBscript will read a template in excel, make a runtime copy of it, Execute a PLSQL procedure in Database, and extract the data from the final tables from database, and populate it into excel template and save the template as location provided with a new name as defined. This will also attach the report into an automated email and sent to user. Job control and error control are tables in database which the vbscipt writes back for its status